+++ crates/api/src/invert.rs
@@ -1,12 +1,17 @@
+// TODO: Remove in 1.0.0 release.
+#![allow(deprecated)]
+
 //! Defines a basic interface for inverting types, similar to that of a fraction.
 //!
 
+#[deprecated(since = "1.0.0", note = "Please use num_traits::Inv")]
 pub trait Invert {
     /// Updates `self` to be inverted.
     ///
     fn invert(self);
 }
 
+#[deprecated(since = "1.0.0", note = "Please use num_traits::Inv")]
 pub trait ToInverse {
     type Output;
 
@@ -15,6 +20,7 @@ pub trait ToInverse {
     fn to_inverse(&self) -> Self::Output;
 }
 
+#[deprecated(since = "1.0.0", note = "Please use num_traits::Inv")]
 pub trait IntoInverse: ToInverse {
     /// Builds a new `Self` that is inverted.
     ///