+++ crates/api/src/parser/terms/mapper/ast_term.rs
@@ -1,6 +1,9 @@
-use super::{Component, Error, Finishable, Term, Visit};
-use crate::{invert::Invert, parser::terms::term_parser::Rule};
+use num_traits::Inv;
 use pest::iterators::Pair;
+
+use crate::parser::terms::term_parser::Rule;
+
+use super::{Component, Error, Finishable, Term, Visit};
 
 pub(super) struct AstTerm {
     pub(super) component: Component,
@@ -41,7 +44,7 @@ impl AstTerm {
                     match op {
                         SecondToken::Dot => (),
                         SecondToken::Slash => {
-                            new_terms.invert();
+                            crate::parser::term::num_traits::inv::inv_terms(&mut new_terms);
                         }
                     }