+++ crates/api/src/unit/deref.rs
@@ -3,6 +3,7 @@ use crate::{parser::Term, unit::Unit};
 //-----------------------------------------------------------------------------
 // impl Deref
 //-----------------------------------------------------------------------------
+// TODO: Get rid of this in 1.0.
 impl ::std::ops::Deref for Unit {
     type Target = [Term];